int64

+++ blender/RecipeComponent.go
@@ -3,7 +3,7 @@ package blender
 type RecipeComponent struct {
 	Component_name   string
 	Component_amount float64
-	Component_id     int
+	Component_id     int64
 	Add_order        int
 	// Lot_id     int64
 	// Product_name_customer_id nullable.NullInt64
